chore: create a core folder and move in the bootstrap file
