+++ src/core/bootstrap.test.js
